Add auth to api routes

+++ pages/api/book/[bookid]/delete.ts
@@ -1,6 +1,7 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from "next";
 import { PrismaClient } from "@prisma/client";
+import { getSession } from "next-auth/react";
 
 const prisma = new PrismaClient();
 
@@ -8,15 +9,22 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  const { bookid } = req.query;
-  try {
-    await prisma.book.delete({
-      where: {
-        id: parseInt(bookid.toString()),
-      },
-    });
-    res.redirect("/");
-  } catch (error) {
-    res.status(403).json({ err: "Error occured while deleting a book item." });
+  const session = await getSession({ req });
+  if (session) {
+    const { bookid } = req.query;
+    try {
+      await prisma.book.delete({
+        where: {
+          id: parseInt(bookid.toString()),
+        },
+      });
+      res.redirect("/");
+    } catch (error) {
+      res
+        .status(403)
+        .json({ err: "Error occured while deleting a book item." });
+    }
+  } else {
+    res.send({ error: "You are not logged in." });
   }
 }
